requirements.txt
asgiref==3.8.1
boto3==1.34.114
botocore==1.34.114
Django==4.2.3
django-cors-headers==4.3.1
django-crequest==2018.5.11
djangorestframework==3.15.1
et-xmlfile==1.1.0
jmespath==1.0.1
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pdfkit==1.0.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
